% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_plot.R
\name{quick_dist}
\alias{quick_dist}
\title{quick distribution plotting of variables in a data frame}
\usage{
quick_dist(df, var_name, na.rm = TRUE)
}
\arguments{
\item{df}{A data frame object. The output plot is based on one of df's variables. Since it should be a data frame, we have called it "df," the initials of the data frame.}

\item{var_name}{A variable (column) in characters format in the provided data frame, df.
It should be an object of numeric, factor (categorical), or character classes.
If numeric, the output is the density plot of the distribution.
If factor or character, the output is a bar chart of frequencies of categories. Since it reflects the name of a variable in characters (double quotations), we have called it "var_name" in the function's body.}

\item{na.rm}{If TRUE, the default, missing values are silently removed. If FALSE, missing values are removed with a warning. We have got the name by putting together the words NA and remove.}
}
\value{
A plotted distribution based on the given column in the data frame input. Based on the class of input variable, it is either a density plot of the distribution or a bar chart of counts of categories.
}
\description{
\code{quick_dist()} function takes two arguments as the input, a data frame and one of its columns.
It plots the distribution of the given variable in the dataset.
}
\examples{
# examples with the gapminder dataset
quick_dist(gapminder::gapminder, "lifeExp")
quick_dist(gapminder::gapminder, "gdpPercap")
quick_dist(gapminder::gapminder, "continent")
}
